backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

minio==7.2.5
aiofiles==23.2.1

redis==5.0.1

python-dotenv==1.0.0

pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
pytest-cov==4.1.0

ruff==0.2.0
black==24.1.1

python-multipart==0.0.6